requirements.txt
Appium_Python_Client==4.5.0
Appium_Python_Client==4.5.0
gradio_client==1.7.0
jsonschema==4.23.0
openai==1.61.1
Pillow==11.1.0
python-slugify==8.0.4
